+++ docs/release-notes/change-log.md
@@ -9,11 +9,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## v1.18.3
 
 ### Server
 
 * Add optional 'secret key' authentication between tier1 and tier2 services
+* Fixed a case where a nil pointer exception could happen on storage error(s).
 * substreams client library will no longer forcefully strip authentication from plaintext connections
 * substreams tier1 will now retry failed tier2 jobs that stream data directly if they have not produced any data yet.
 * substreams worker jobs will always retry tier2 jobs that return with codes.Unavailable: `no healthy upstream`, considering this error as load-balancer version of codes.ResourceExhausted